Fix issue #594 Data integrity of inputs through to outputs (ie. raw, Channels, ppmChannels) Type s16 is not atomic value, the CPU uses two instructions to copy 32-bit register value first and then set hi bytes to zero during the assignment. There's chances that the CPU is interrupted to service the protocol callbacks between the two instructions. To ensure data integrity, consider changing the array type to s32. It take an extra 20 bytes, but the array elements are atomic values, can never have middle value to be read.

+++ src/common.h
@@ -34,7 +34,7 @@ struct FAT {
 };
 
 
-extern volatile s16 Channels[NUM_OUT_CHANNELS];
+extern volatile s32 Channels[NUM_OUT_CHANNELS];
 extern const char DeviationVersion[33];
 
 /* Temproary definition until we have real translation */
