+++ src/pages/128x64x1/dialogs.c
@@ -35,14 +35,14 @@ static const char *safety_string_cb(guiObject_t *obj, void *data)
     int i;
     int count = 0;
     const s8 safeval[4] = {0, -100, 0, 100};
-    volatile s16 *raw = MIXER_GetInputs();
+    volatile s32 *raw = MIXER_GetInputs();
     tempstring[0] = 0;
     for(i = 0; i < NUM_SOURCES + 1; i++) {
         if (! (unsafe & (1LL << i)))
             continue;
         int ch = (i == 0) ? PROTOCOL_MapChannel(INP_THROTTLE, NUM_INPUTS + 2) : i-1;
 
-        s16 val = RANGE_TO_PCT((ch < NUM_INPUTS)
+        s32 val = RANGE_TO_PCT((ch < NUM_INPUTS)
                       ? raw[ch+1]
                       : MIXER_GetChannel(ch - (NUM_INPUTS), APPLY_SAFETY));
         INPUT_SourceName(tempstring + strlen(tempstring), ch + 1);
